+++ src/main/java/chess/domain/game/type/Command.java
@@ -1,4 +1,4 @@
-package chess.controller;
+package chess.domain.game.type;
 
 import java.util.Arrays;
