+++ src/main/java/chess/view/OutputView.java
@@ -3,8 +3,8 @@ package chess.view;
 import chess.controller.dto.ScoresDTO;
 import chess.domain.board.Board;
 import chess.domain.board.Cell;
-import chess.domain.board.File;
-import chess.domain.board.Rank;
+import chess.domain.board.type.File;
+import chess.domain.board.type.Rank;
 import chess.domain.piece.Piece;
 import chess.domain.player.type.TeamColor;
 import chess.domain.position.Position;
