+++ src/test/java/chess/domain/board/BoardTest.java
@@ -12,7 +12,6 @@ import chess.domain.position.MoveRoute;
 import chess.domain.position.Position;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 class BoardTest {
@@ -23,51 +22,47 @@ class BoardTest {
         board = new ChessGame(new BoardDefaultSetting()).board();
     }
 
+    @DisplayName("출발 위치에 자신의 기물이 없는 경우 이동 불가 - 빈 칸인 경우")
+    @Test
+    void cannotMovePieceAtStartPositionEmpty() {
+        MoveRoute moveRoute = new MoveRoute("a3", "a4");
+        MoveCommand moveCommand = new MoveCommand(WHITE, moveRoute);
+
+        assertThatThrownBy(() -> board.move(moveCommand))
+            .isInstanceOf(IllegalArgumentException.class);
+    }
+
+    @DisplayName("출발 위치에 자신의 기물이 없는 경우이동 불가 - 적의 기물이 있는 경우")
+    @Test
+    void cannotMovePieceAtStartPositionEnemyPiece() {
+        MoveRoute moveRoute = new MoveRoute("a7", "a6");
+        MoveCommand moveCommand = new MoveCommand(WHITE, moveRoute);
+
+        assertThatThrownBy(() -> board.move(moveCommand))
+            .isInstanceOf(IllegalArgumentException.class);
+    }
+
     @DisplayName("기물 이동")
-    @Nested
-    class MovePiece {
-        @DisplayName("출발 위치에 자신의 기물이 없는 경우 이동 불가 - 빈 칸인 경우")
-        @Test
-        void cannotMovePieceAtStartPositionEmpty() {
-            MoveRoute moveRoute = new MoveRoute("a3", "a4");
-            MoveCommand moveCommand = new MoveCommand(WHITE, moveRoute);
+    @Test
+    void movePiece() {
+        MoveRoute moveRoute = new MoveRoute("a2", "a4");
+        MoveCommand moveCommand = new MoveCommand(WHITE, moveRoute);
+        board.move(moveCommand);
 
-            assertThatThrownBy(() -> board.move(moveCommand))
-                .isInstanceOf(IllegalArgumentException.class);
-        }
+        assertThat(board.findCell(Position.of("a2")).isEmpty()).isTrue();
+        assertThat(board.findPiece(Position.of("a4"))).isEqualTo(new Pawn(WHITE));
+    }
 
-        @DisplayName("출발 위치에 자신의 기물이 없는 경우이동 불가 - 적의 기물이 있는 경우")
-        @Test
-        void cannotMovePieceAtStartPositionEnemyPiece() {
-            MoveRoute moveRoute = new MoveRoute("a7", "a6");
-            MoveCommand moveCommand = new MoveCommand(WHITE, moveRoute);
+    @DisplayName("기물이 이동할 수 없는 도착위치")
+    @Test
+    void cannotMovePieceToDestination() {
+        MoveRoute moveRoute = new MoveRoute("a2", "a5");
+        MoveCommand moveCommand = new MoveCommand(WHITE, moveRoute);
 
-            assertThatThrownBy(() -> board.move(moveCommand))
-                .isInstanceOf(IllegalArgumentException.class);
-        }
+        assertThatThrownBy(() -> board.move(moveCommand))
+            .isInstanceOf(IllegalArgumentException.class);
 
-        @DisplayName("기물 이동")
-        @Test
-        void movePiece() {
-            MoveRoute moveRoute = new MoveRoute("a2", "a4");
-            MoveCommand moveCommand = new MoveCommand(WHITE, moveRoute);
-            board.move(moveCommand);
-
-            assertThat(board.findCell(Position.of("a2")).isEmpty()).isTrue();
-            assertThat(board.findPiece(Position.of("a4"))).isEqualTo(new Pawn(WHITE));
-        }
-
-        @DisplayName("기물이 이동할 수 없는 도착위치")
-        @Test
-        void cannotMovePieceToDestination() {
-            MoveRoute moveRoute = new MoveRoute("a2", "a5");
-            MoveCommand moveCommand = new MoveCommand(WHITE, moveRoute);
-
-            assertThatThrownBy(() -> board.move(moveCommand))
-                .isInstanceOf(IllegalArgumentException.class);
-
-            assertThat(board.findPiece(Position.of("a2"))).isEqualTo(new Pawn(WHITE));
-            assertThat(board.findCell(Position.of("a5")).isEmpty()).isTrue();
-        }
+        assertThat(board.findPiece(Position.of("a2"))).isEqualTo(new Pawn(WHITE));
+        assertThat(board.findCell(Position.of("a5")).isEmpty()).isTrue();
     }
 }
