+++ src/test/java/chess/domain/position/MoveRouteTest.java
@@ -1,9 +1,9 @@
 package chess.domain.position;
 
-import static chess.domain.board.File.A;
-import static chess.domain.board.File.C;
-import static chess.domain.board.Rank.FIVE;
-import static chess.domain.board.Rank.THREE;
+import static chess.domain.board.type.File.A;
+import static chess.domain.board.type.File.C;
+import static chess.domain.board.type.Rank.FIVE;
+import static chess.domain.board.type.Rank.THREE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
